Module jack::client::client_status

Contains ClientStatus flags which describe the status of a Client.

Structs

ClientStatus
Status flags for JACK clients.

Constants

BACKEND_ERROR
No documentation found. TODO: dig deeper

CLIENT_ZOMBIE
No documentation found. TODO: dig deeper

FAILURE
Overall operation failed. File an issue if you can get it to appear.

INIT_FAILURE
Unable to initialize client

INVALID_OPTION
The operation contained an invalid or unsupported option.

LOAD_FAILURE
Unable to load internal client

NAME_NOT_UNIQUE
The desired client name was not unique. With the USE_EXACT_NAME option this situation is fatal. Otherwise, the name was modified by appending a dash and a two-digit number in the range "-01" to "-99". Client::name() will return the exact string that was used. If the specified client_name plus these extra characters would be too long, the open fails instead.

NO_SUCH_CLIENT
Requested client does not exist.

SERVER_ERROR
Communication error with the JACK server.

SERVER_FAILED
Unable to connect to the JACK server.

SERVER_STARTED
The JACK server was started as a result of this operation. Otherwise, it was running already. In either case the caller is now connected to jackd, so there is no race condition. When the server shuts down, the client will find out.

SHM_FAILURE
Unable to access shared memory

VERSION_ERROR
Client's protocol version does not match