Module jack::logging

Control error and info logging from JACK.

Functions

get_error_callback
Get the error callback that was set using set_error_callback. This corresponds to the one set using rust-jack, not JACK itself. None is returned if rust-jack hasn't set a callback or has reset it to use stderr.

get_info_callback
Resets the JACK info callback to use stdio. Get the info callback that was set using set_info_callback. This corresponds to the one set using rust-jack, not JACK itself. None is returned if rust-jack hasn't set a callback or has reset it to use stdout.

reset_error_callback
Restores the JACK info callback to the JACK default, which is to write to stderr.

reset_info_callback
Restores the JACK info callback to the JACK default, which is to write to stdout.

set_error_callback
Set the global JACK error callback. It is recommended to use the log crate.

set_info_callback
Set the global JACK info callback. It is recommended to use the log crate.