Module jack::port::port_flags

Contains flag constants that may be used to create PortFlags.

Structs

PortFlags
Flags for specifying port options.

Constants

CAN_MONITOR
A call to jack_port_request_monitor() makes sense.

IS_INPUT
The port can receive data.

IS_OUTPUT
Data can be read from the port.

IS_PHYSICAL
Port corresponds to some kind of physical I/O connector.

IS_TERMINAL
For an input port, the data received by the port will not be passed on or made available at any other port. For output, the data available at the port does not originate from any other port. Audio synthesizers, I/O hardware interface clients, HDR systems are examples of clients that would set this flag for their ports.