Struct jack::port::Unowned

pub struct Unowned;

PortSpec for a port that holds has no readable or writeable data from JACK on the created client. It can be used for obtaining information about external ports.