Module jack::prelude

Contains most functionality needed to interact with JACK.

Reexports

pub use primitive_types::JackFrames;

pub use primitive_types::JackTime;

pub use primitive_types::JackPortId;

pub use jack_enums::JackErr;

pub use jack_enums::JackControl;

pub use jack_enums::LatencyType;

pub use client::ClientOptions;

pub use client::ClientStatus;

pub use client::client_options;

pub use client::client_status;

pub use ringbuffer::RingBuffer;

pub use ringbuffer::RingBufferReader;

pub use ringbuffer::RingBufferWriter;

pub use port::PortFlags;

pub use port::port_flags;

pub use logging::set_info_callback;

pub use logging::get_info_callback;

pub use logging::reset_info_callback;

pub use logging::set_error_callback;

pub use logging::get_error_callback;

pub use logging::reset_error_callback;

Structs

AsyncClient
A JACK client that is processing data asynchronously, in real-time.

AudioInPort
Safely and thinly wrap a Port<AudioInSpec>. Derefs into a &[f32].

AudioInSpec
AudioInSpec implements the PortSpec trait which, defines an endpoint for JACK. In this case, it is a readable 32 bit floating point buffer for audio.

AudioOutPort
Safely and thinly wrap a Port<AudioOutSpec>. Derefs into a &mut[f32].

AudioOutSpec
AudioOutSpec implements the PortSpec trait, which defines an endpoint for JACK. In this case, it is a mutable 32 bit floating point buffer for audio.

CLIENT_NAME_SIZE
The maximum string length for port names.

Client
A client to interact with a JACK server.

ClosureProcessHandler
Wrap a closure that can handle the process callback. This is called every time data from ports is available from JACK.

CycleTimes
Internal cycle timing information.

MidiInPort
Safely and thinly wrap a Port<MidiInPort>.

MidiInSpec
MidiInSpec implements the PortSpec trait, which defines an endpoint for JACK. In this case, it defines midi input.

MidiIter
Iterate through Midi Messages within a MidiInPort.

MidiOutPort
Safely and thinly wrap a Port<MidiInPort>.

MidiOutSpec
MidiOutSpec implements the PortSpec trait, which defines an endpoint for JACK. In this case, it defines a midi output.

PORT_NAME_SIZE
The maximum string length for port names.

PORT_TYPE_SIZE
The maximum string length for jack type names.

Port
An endpoint to interact with JACK data streams, for audio, midi, etc...

ProcessScope
ProcessScope provides information on the client and frame time information within a process callback.

RawMidi
Contains 8bit raw midi information along with a timestamp relative to the process cycle.

Unowned
PortSpec for a port that holds has no readable or writeable data from JACK on the created client. It can be used for obtaining information about external ports.

Traits

NotificationHandler
Specifies callbacks for JACK.

PortSpec
Defines the configuration for a certain port to JACK, ie 32 bit floating audio input, 8 bit raw midi output, etc...

ProcessHandler

Type Definitions

UnownedPort
Port<UnownedSpec> - Port that holds no data from Jack, though it can still be used to query information.