Struct jack::prelude::ClosureProcessHandler

pub struct ClosureProcessHandler<F: 'static + Send + FnMut(&Client, &ProcessScope) -> JackControl> {
    pub process_fn: F,
}

Wrap a closure that can handle the process callback. This is called every time data from ports is available from JACK.

Fields
process_fn: F
Methods

impl<F> ClosureProcessHandler<F> where
F: 'static + Send + FnMut(&Client, &ProcessScope) -> JackControl,

fn new(f: F) -> ClosureProcessHandler<F>