Struct jack::prelude::MidiOutPort

pub struct MidiOutPort<'a> { /* fields omitted */ }

Safely and thinly wrap a Port<MidiInPort>.

Methods

impl<'a> MidiOutPort<'a>

fn new(port: &'a mut Port<MidiOutSpec>, ps: &'a ProcessScope) -> Self

Wrap a Port<MidiInSpec> within a process scope of a client that registered the port. Panics if the port does not belong to the client that created the process.

The data in the port is cleared.

fn write(&mut self, message: &RawMidi) -> Result<(), JackErr>

Write an event into an event port buffer.

Clients must write normalised MIDI data to the port - no running status and no (1-byte) realtime messages interspersed with other messagse (realtime messages are fine when they occur on their own, like other messages).

fn lost_count(&self) -> usize

Get the number of events that could not be written to port_buffer.

If the return value is greater than 0, than the buffer is full. Currently, the only way this can happen is if events are lost on port mixdown.

fn max_event_size(&self) -> usize

Get the size of the largest event that can be stored by the port.

This function returns the current space available, taking into account events already stored in the port.