Struct jack::ringbuffer::RingBufferReader

pub struct RingBufferReader { /* fields omitted */ }

Read end of the ring buffer. Can only be used from one thread (can be different from the write thread).

Methods

impl RingBufferReader

fn get_vector<'a>(&'a self) -> (&'a [u8], &'a [u8])

Fill a data structure with a description of the current readable data held in the ringbuffer. This description is returned in a two slices. Two slices are needed because the data to be read may be split across the end of the ringbuffer. The first slice represents the bytes ready to be read. if the second slice

fn read_buffer(&mut self, buf: &mut [u8]) -> usize

Read data from the ringbuffer. Returns: the number of bytes read, which may range from 0 to buf.len().

fn peek(&self, buf: &mut [u8]) -> usize

Read data from the ringbuffer. Opposed to read_buffer() this function does not move the read pointer. Thus it's a convenient way to inspect data in the ringbuffer in a continous fashion. The price is that the data is copied into a user provided buffer. For "raw" non-copy inspection of the data in the ringbuffer use get_vector() or peek_iter. Returns: the number of bytes read, which may range from 0 to buf.len()

fn advance(&mut self, cnt: usize)

Advance the read pointer. use this after peek/peek_iter or get_vector to advance the buffer pointer.

fn space(&self) -> usize

Return the number of bytes available for reading.

fn peek_iter<'a>(&'a self) -> Chain<Iter<'a, u8>, Iter<'a, u8>>

Iterator that goes over all the data available to read.