Struct jack::MidiWriter

pub struct MidiWriter<'a> { /* fields omitted */ }

Write midi events to an output midi port.

Methods

impl<'a> MidiWriter<'a>

pub fn write(&mut self, message: &RawMidi) -> Result<(), Error>

Write an event into an event port buffer.

Clients must write normalised MIDI data to the port - no running status and no (1-byte) realtime messages interspersed with other messagse (realtime messages are fine when they occur on their own, like other messages).

pub fn lost_count(&self) -> usize

Get the number of events that could not be written to port_buffer.

If the return value is greater than 0, than the buffer is full. Currently, the only way this can happen is if events are lost on port mixdown.

pub fn max_event_size(&self) -> usize

Get the size of the largest event that can be stored by the port.

This function returns the current space available, taking into account events already stored in the port.