Struct jack::Port

pub struct Port<PS> { /* fields omitted */ }

An endpoint to interact with JACK data streams, for audio, midi, etc...

The Port struct contains mostly metadata and exposes data as raw pointers. For a better data consumption/production API, see the AudioInPort, AudioOutPort, MidiInPort, and MidiOutPort.

Most JACK functionality is exposed, including the raw pointers, but it should be possible to create a client without the need for calling unsafe Port methods.

Methods

impl Port<AudioIn>

pub fn as_slice<'a>(&'a self, ps: &'a ProcessScope) -> &'a [f32]

Read the received audio data.

impl Port<AudioOut>

pub fn as_mut_slice<'a>(&'a mut self, ps: &'a ProcessScope) -> &'a mut [f32]

Get a slice to write audio data to.

impl Port<MidiIn>

pub fn iter<'a>(&'a self, ps: &'a ProcessScope) -> MidiIter<'a>

Get an iterator over midi events.

impl Port<MidiOut>

pub fn writer<'a>(&'a mut self, ps: &'a ProcessScope) -> MidiWriter<'a>

Create a writer that can write midi events to the specified midi port. Calling this function clears the midi buffer.

impl<PS> Port<PS>

pub fn spec(&self) -> &PS

Returns the spec that was used to create this port.

pub fn clone_unowned(&self) -> Port<Unowned>

Return a copy of port as an unowned port that can still be used for querying information.

pub fn name<'a>(&'a self) -> &'a str

Returns the full name of the port, including the "client_name:" prefix.

pub fn short_name<'a>(&'a self) -> &'a str

Returns the short name of the port, it excludes the "client_name:" prefix.

pub fn flags(&self) -> PortFlags

The flags for the port. These are set when the port is registered with its client.

pub fn port_type<'a>(&self) -> &'a str

The port type. JACK's built in types include "32 bit float mono audio" and "8 bit raw midi". Custom types may also be used.

pub fn connected_count(&self) -> usize

Number of ports connected to/from &self.

pub fn is_connected_to(&self, port_name: &str) -> bool

Returns true if the port is directly connected to a port with the name port_name.

pub fn aliases(&self) -> Vec<String>

Get the alias names for self.

Will return up to 2 strings.

pub fn is_monitoring_input(&self) -> bool

Returns true if monitoring has been requested for self.

pub fn request_monitor(&self, enable_monitor: bool) -> Result<(), Error>

Turn input monitoring for the port on or off.

This only works if the port has the CAN_MONITOR flag set.

pub fn ensure_monitor(&self, enable_monitor: bool) -> Result<(), Error>

If the CAN_MONITOR flag is set for the port, then input monitoring is turned on if it was off, and turns it off if only one request has been made to turn it on. Otherwise it does nothing.

pub fn set_name(&mut self, short_name: &str) -> Result<(), Error>

Set's the short name of the port. If the full name is longer than PORT_NAME_SIZE, then it will be truncated.

pub fn set_alias(&mut self, alias: &str) -> Result<(), Error>

Sets alias as an alias for self.

May be called at any time. If the alias is longer than PORT_NAME_SIZE, it will be truncated.

After a successful call, and until JACK exists, or the alias is unset, alias may be used as an alternate name for the port.

Ports can have up to two aliases - if both are already set, this function will return an error.

pub fn unset_alias(&mut self, alias: &str) -> Result<(), Error>

Remove alias as an alias for port. May be called at any time.

After a successful call, alias can no longer be used as an alternate name for self.

pub unsafe fn from_raw(
spec: PS,
client_ptr: *mut jack_client_t,
port_ptr: *mut jack_port_t
) -> Self

Create a Port from raw JACK pointers.

This is mostly for use within the jack crate itself.

pub fn client_ptr(&self) -> *mut jack_client_t

Obtain the client pointer that spawned this port.

This is mostly for use within the jack crate itself.

pub fn raw(&self) -> *mut jack_port_t

Obtain the ffi port pointer.

This is mostly for use within the jack crate itself.

pub unsafe fn buffer(&self, n_frames: Frames) -> *mut c_void

Obtain the buffer that the Port is holding. For standard audio and midi ports, consider using the AudioInPort, AudioOutPort, MidiInPort, or MidiOutPort adapter. For more custom data, consider implementing your own adapter that safely uses the Port::buffer method.